add token endpoint api support

+++ demo/sdk_demo_v2/http_client.h
@@ -158,7 +158,7 @@ namespace saz {
 					body.append(buffer, buffer + read_size);
 				}
 
-				return {status_code, body};
+				return { static_cast<int>(status_code), body };
 			}
 
 		public:
